requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
Flask==2.2.2
pymongo==4.3.3
Werkzeug==2.2.2
requests
PyPDF2